[client] extract communities.py out of bot-community.py
